% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Calc_derived_quants}
\alias{Calc_derived_quants}
\title{Copy of derive_results}
\usage{
Calc_derived_quants(...)
}
\description{
Included for continuity when using old scripts
}
\details{
Please use \code{?derive_results} to see list of arguments and usage
}
