% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCSRA.R
\docType{package}
\name{CCSRA}
\alias{CCSRA}
\alias{CCSRA-package}
\title{Catch curve stock reduction analysis}
\description{
CCSRA is an age-structured stock assessment modelling package built in TMB
}
\details{
CCSRA is desired to implement catch-curve analysis, stock-reduction analysis, age-structured models, age-structured production models, and catch-curve stock reduction analysis.  It is intended to facilitate learning and simulation-testing across multiple model types, particularly when comparing mixed-effect and penalized-likelihood implementations.

See examples folder on github \url{http://github.com/james-thorson/CCSRA/examples} for up-to-date examples using lower-level functions.
}
