% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_length_to_age_samples.R
\name{convert_length_to_age_samples}
\alias{convert_length_to_age_samples}
\title{Convert length-composition to age-composition data}
\usage{
convert_length_to_age_samples(K, Linf, L0, Lcv, Lbin_mat, LengthComp_lt,
  checkforbugs = TRUE)
}
\arguments{
\item{K, }{Brody growth coefficient}

\item{Linf, }{Asymptotic maximum length}

\item{L0, }{Expected length at age 0 from fitting a von Bertalanffy growth curve}

\item{Lcv, }{coefficient of variation for size given expected length at age}

\item{Lbin_mat, }{a matrix with two columns, where each row specifies the lower and upper length for a given length bin (the lowest should be -Inf, and highest Inf)}

\item{LengthComp_lt, }{a matrix of length-composition data, where columns are samples in year t, and cells are the count of samples with a given length and year}

\item{checkforbugs, }{a boolean stating whether to check for bugs or not}
}
\value{
AgeComp_at, a matrix of expected age-composition data, where columns are samples in year t, and cells are the count of samples with a given age and year
}
\description{
\code{convert_length_to_age_samples} converts a matrix of length-composition data to age-composition data for use in CCSRA
}
