% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_outputs.R
\name{derive_outputs}
\alias{derive_outputs}
\title{Calculate biological reference points}
\usage{
derive_outputs(Obj, SD = NULL)
}
\arguments{
\item{Obj, }{the fitted TMB object}
}
\value{
List, a tagged list of potentially useful benchmarks
}
\description{
\code{derive_outputs} calculates derived quanties for status or productivity
}
