% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_results.R
\name{derive_results}
\alias{derive_results}
\title{Calculate biological reference points}
\usage{
derive_results(Obj, SD = NULL)
}
\arguments{
\item{Obj, }{the fitted TMB object}
}
\value{
List, a tagged list of potentially useful benchmarks
}
\description{
\code{derive_results} calculates derived quanties for status or productivity
}
