% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{FormatInput_Fn}
\alias{FormatInput_Fn}
\title{Copy of make_inputs}
\usage{
FormatInput_Fn(...)
}
\description{
Included for continuity when using old scripts
}
\details{
Please use \code{?make_inputs} to see list of arguments and usage
}
