% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_inputs.R
\name{make_inputs}
\alias{make_inputs}
\title{Build inputs for CCSRA}
\usage{
make_inputs(Version = "CCSRA_v8", Method, M_prior, h_prior, D_prior,
  SigmaR_prior, Sslope_prior = c(-999, 999, 1), AgeComp_at, Index_t,
  Cw_t, W_a, Mat_a, RecDev_biasadj, rec_method = "dev",
  estimate_recdevs = TRUE, use_dirmult = FALSE)
}
\value{
a list with inputs for building CCSRA TMB object
}
\description{
\code{make_inputs} builds inputs necessary for running CCSRA
}
