% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{SimData_Fn}
\alias{SimData_Fn}
\title{Copy of simulate_data}
\usage{
SimData_Fn(...)
}
\description{
Included for continuity when using old scripts
}
\details{
Please use \code{?simulate_data} to see list of arguments and usage
}
