% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Convert_length_to_age_samples.R
\name{Convert_length_to_age_samples}
\alias{Convert_length_to_age_samples}
\title{Convert length-composition to age-composition data}
\usage{
Convert_length_to_age_samples(K, Linf, L0, Lcv, Lbin_mat, AgeComp_lt,
  checkforbugs = TRUE)
}
\arguments{
\item{K,}{Brody growth coefficient}

\item{Linf,}{Asymptotic maximum length}

\item{L0,}{Expected length at age 0 from fitting a von Bertalanffy growth curve}

\item{Lcv,}{coefficient of variation for size given expected length at age}

\item{Lbin_mat,}{a matrix with two columns, where each row specifies the lower and upper length for a given length bin (the lowest should be -Inf, and highest Inf)}

\item{AgeComp_lt,}{a matrix of length-composition data, where columns are samples in year t, and cells are the count of samples with a given length and year}

\item{checkforbugs,}{a boolean stating whether to check for bugs or not}
}
\value{
AgeComp_at, a matrix of expected age-composition data, where columns are samples in year t, and cells are the count of samples with a given age and year
}
\description{
\code{Convert_length_to_age_samples} converts a matrix of length-composition data to age-composition data for use in CCSRA
}

