% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate data for fitting CCSRA}
\usage{
simulate_data(
  Nyears,
  AgeMax,
  SigmaR,
  M,
  F1,
  W_a,
  S_a,
  Mat_a,
  h,
  SB0,
  R0,
  Frate,
  Fequil,
  SigmaF,
  F_method = 1,
  Fdynamics = "Endogenous",
  Fmax = NA,
  Ncomp_per_year,
  SurveyCV,
  Recruitment_dynamics = "Stationary",
  Regime_multiplier = NA,
  SigmaS = 0,
  rhoA = 0.8,
  rhoT = 0.8
)
}
\description{
\code{simulate_data} simulates data for use in testing CCSRA
}
